test(gateway): extract session hub unit tests
